fix license
