+++ be/CMakeLists.txt
@@ -914,6 +914,7 @@ add_subdirectory(${SRC_DIR}/exprs)
 add_subdirectory(${SRC_DIR}/format)
 add_subdirectory(${SRC_DIR}/gen_cpp)
 add_subdirectory(${SRC_DIR}/io)
+add_subdirectory(${SRC_DIR}/vec/sink/writer/paimon/compat)
 add_subdirectory(${SRC_DIR}/storage/index/ann)
 add_subdirectory(${SRC_DIR}/storage)
 add_subdirectory(${SRC_DIR}/runtime)
