+++ be/src/vec/sink/writer/paimon/compat/rawlog_compat.cpp
@@ -1,0 +1,39 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+#include <glog/logging.h>
+
+#include <cstdarg>
+#include <cstdio>
+
+namespace google {
+void RawLog__(LogSeverity severity, const char* file, int line, const char* format, ...) {
+    char buffer[2048];
+    va_list ap;
+    va_start(ap, format);
+    std::vsnprintf(buffer, sizeof(buffer), format, ap);
+    va_end(ap);
+    const char* sev = "INFO";
+    if (severity == GLOG_WARNING)
+        sev = "WARN";
+    else if (severity == GLOG_ERROR)
+        sev = "ERROR";
+    else if (severity == GLOG_FATAL)
+        sev = "FATAL";
+    std::fprintf(stderr, "[%s] %s:%d %s\n", sev, file ? file : "", line, buffer);
+}
+} // namespace google
